requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
PyMySQL==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
cryptography==41.0.7
gunicorn==21.2.0
requests==2.31.0